changing the packages (#10)
* changing the packages

* moving VRepotTests

* reportRow name fix

* update doc

* Delete UIChromeTestBase.kt

* Delete UITestBase.kt
