+++ galite-core/src/main/kotlin/org/kopi/galite/chart/Chart.kt
@@ -15,7 +15,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package org.kopi.galite.visual.chart
+package org.kopi.galite.chart
 
 import org.kopi.galite.domain.Domain
