+++ galite-core/src/main/kotlin/org/kopi/galite/exceptions/InvalidValueException.kt
@@ -15,33 +15,15 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package org.kopi.galite.visual.chart
-
-import org.kopi.galite.domain.Domain
-import org.kopi.galite.visual.field.Field
+package org.kopi.galite.exceptions
 
 /**
- * Represents a one dimension that contains measures [values] to use in chart.
+ * Thrown to indicate that an invalid value has been passed to a field.
  *
- * @param domain dimension domain.
+ * @param value to pass
+ * @param label the field's label
+ *
  */
-open class Dimension<T : Comparable<T>>(domain: Domain<T>? = null) : Field<T>(domain) {
-
-  /**
-   * Dimension values
-   */
-  val values = mutableListOf<DimensionData<T>>()
-
-  /**
-   * Add a dimension value
-   *
-   * @param value the dimension value
-   */
-  fun add(value: T, init: (DimensionData<T>.() -> Unit)? = null) {
-    val dimensionValue: DimensionData<T> = DimensionData<T>(value)
-    if (init != null) {
-      dimensionValue.init()
-    }
-    values.add(dimensionValue)
-  }
+class InvalidValueException(value: Comparable<*>, label: String) : RuntimeException() {
+  override val message = "invalid value $value for the field $label"
 }
