+++ galite-core/src/main/kotlin/org/kopi/galite/report/ReportRow.kt
@@ -1,0 +1,51 @@
+/*
+ * Copyright (c) 2013-2020 kopiLeft Services SARL, Tunis TN
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License version 2.1 as published by the Free Software Foundation.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+
+package org.kopi.galite.report
+
+import org.kopi.galite.field.Field
+
+/**
+ * Represents a data row of a [Report].
+ *
+ * @param reportFields the fields that exists in the report.
+ */
+class ReportRow(private val reportFields: MutableList<Field<*>>) {
+  /** A report data row */
+  val data = mutableMapOf<Field<*>, Any>()
+
+  /**
+   * Returns data value for a specific [field].
+   *
+   * @param field the field.
+   * @return  data value for a specific [field].
+   */
+  fun getValueOf(field: Field<*>) = data[field]
+
+  /**
+   * Sets a mapping between the values that the domain can take
+   * and a corresponding text to be displayed in a [Field].
+   *
+   * @param field the field.
+   * @param value the field's value.
+   */
+  operator fun <T: Comparable<T>> set(field: Field<T>, value: T) {
+    if(field in reportFields) {
+      data.putIfAbsent(field, value)
+    }
+  }
+}
